moved the parent form and view classes into a generics module to make the normal forms and views files easier to read by example
