+++ contact/templates/default_form.html
@@ -11,7 +11,7 @@
           {% csrf_token %}
 
           {% for field in form.hidden_fields %}
-            {% include "forms/hiddenfield.html" %}
+            {% include "includes/hiddenfield.html" %}
           {% endfor %}
 
           {% for field in form.visible_fields %}
