+++ contact/templates/default_thanks.html
@@ -8,7 +8,7 @@
         Sorry, your request has failed - code {{ success_data.code }}
     {% endif %}
     {% if success_data.originating_page %}
-        <a href="{{ originating_page }}">Click here to go back</a>
+        <a href="{{ success_data.originating_page }}">Click here to go back</a>
     {% endif %}
   </div>
