+++ contact/templatetags/utils.py
@@ -7,6 +7,6 @@ register = template.Library()
 @register.simple_tag
 def get_form_field_template(boundfield):
     if type(boundfield.field.widget) in [TextInput, EmailInput]:
-        return 'forms/inputfield.html'
+        return 'includes/inputfield.html'
     elif type(boundfield.field.widget) == Textarea:
-        return 'forms/textarea.html'
+        return 'includes/textarea.html'
